references/LLM-Enhanced-Trading/requirements.txt
fastapi
uvicorn[standard]
requests
pandas
numpy
pandas_market_calendars
websocket-client
pytz
praw
cohere
newsapi-python
pydantic
python-multipart
transformers==4.40.1
peft==0.5.0
sentencepiece
accelerate
torch
datasets
bitsandbytes
huggingface-hub
python-dotenv